Refactor test infrastructure with shared utilities and JSDOM setup
